Attempt to fix fullscreen defocus bug on Windows
On Windows, if a borderless fullscreen window last presented a texture with the same dimensions as its current monitor, when it subsequently loses focus, the presented texture remains, obscuring the screen and causing confusing results. This is an attempted fix which simply adds a single pixel to the vertical dimension of any fullscreen window when it is resized.